optimizing message sending, groups & repositories (for better performance) section 17 finished

+++ API/SignalR/MessageHub.cs
@@ -21,14 +21,18 @@ IUserRepository userRepository, IMapper mapper, IHubContext<PresenceHub> presenc
         throw new Exception("Cannot join group");
         var groupName = GetGroupName(Context.User.GetUsername(), otherUser);
         await Groups.AddToGroupAsync(Context.ConnectionId, groupName);
-        await AddToGroup(groupName);
+        var group = await AddToGroup(groupName);
+
+        await Clients.Group(groupName).SendAsync("UpdatedGroup", group);
+
         var messages = await messageRepository.GetMessageThread(Context.User.GetUsername(), otherUser!);
 
-        await Clients.Group(groupName).SendAsync("ReceiveMessageThread", messages);
+        await Clients.Caller.SendAsync("ReceiveMessageThread", messages);
     }
     public override async Task OnDisconnectedAsync(Exception? exception)
     {
-        await RemoveFromMessageGroup();
+        var group = await RemoveFromMessageGroup();
+        await Clients.Group(group.Name).SendAsync("UpdatedGroup", group);
         await base.OnDisconnectedAsync(exception);
     }
     public async Task SendMessage(CreateMessageDto createMessageDto) 
@@ -79,7 +83,7 @@ IUserRepository userRepository, IMapper mapper, IHubContext<PresenceHub> presenc
         }
     }
 
-    private async Task<bool> AddToGroup(string groupName) 
+    private async Task<Group> AddToGroup(string groupName) 
     {
         var username = Context.User?.GetUsername() ?? throw new Exception("Cannot get username!");
         var group = await messageRepository.GetMessageGroup(groupName);
@@ -92,17 +96,21 @@ IUserRepository userRepository, IMapper mapper, IHubContext<PresenceHub> presenc
         }
         group.Connections.Add(connection);
 
-        return await messageRepository.SaveAllAsync();
+        if (await messageRepository.SaveAllAsync()) return group;
+
+        throw new HubException("Failed to join group!");
     }
 
-    private async Task RemoveFromMessageGroup() 
+    private async Task<Group> RemoveFromMessageGroup() 
     {
-        var connection = await messageRepository.GetConnection(Context.ConnectionId);
-        if(connection != null) 
+        var group = await messageRepository.GetGroupForConnection(Context.ConnectionId);
+        var connection = group?.Connections.FirstOrDefault( x=> x.ConnectionId == Context.ConnectionId);
+        if(connection != null && group != null) 
         {
             messageRepository.RemoveConnection(connection);
-            await messageRepository.SaveAllAsync();
+            if (await messageRepository.SaveAllAsync()) return group;
         }
+        throw new Exception("Failed to remove from group!");
     }
 
     private string GetGroupName(string caller, string? other)
